NYT api call and screens done
