+++ src/screens/newYorkTimes/single/_styles.ts
@@ -2,6 +2,7 @@ import {StyleSheet} from 'react-native';
 import {moderateScale, scale, verticalScale} from 'react-native-size-matters';
 //components
 import {LIKE_ICON_SIZE} from '../../../components/LikeIcon';
+import {theme} from '../../../style/styles';
 
 const image = StyleSheet.create({
   image: {
@@ -29,6 +30,15 @@ const details = StyleSheet.create({
   description: {
     marginTop: moderateScale(10),
     textAlign: 'justify',
+  },
+  buyLinksContainer: {
+    borderWidth: moderateScale(1),
+    marginTop: moderateScale(10),
+    alignItems: 'center',
+    height: moderateScale(40),
+    justifyContent: 'center',
+    borderColor: theme.color.darkGray,
+    borderRadius: moderateScale(20),
   },
   buyLinks: {
     marginTop: moderateScale(10),
